// functions/src/stats/analysis/types.ts
export type AnalysisTypeId =
  | "GOAT"
  | "COMPLETE_PLAYER"
  | "ALL_ROUNDER"
  | "TWO_WAY_PLAYER"
  | "DEEP_BAG"
  | "SPARK_PLUG"
  | "BIG_GAME_HUNTER"
  | "SHARPSHOOTER"
  | "CLUTCH"
  | "CHAOS_RUNNER"
  | "WALKING_BUCKET"
  | "HIGH_FLOOR"
  | "CHAOS_ANCHOR"
  | "FINISHER"
  | "LASER"
  | "HIGH_MOTOR"
  | "CHAOS_TAKER"
  | "IRON_MAN"
  | "PROSPECT"
  /** @deprecated 旧判定。月次 V1 では使わない */
  | "CHEAT_CODE"
  | "ELITE_ALLROUNDER"
  | "GIANT_SLAYER"
  | "HOT_HAND"
  | "UNICORN"
  | "ASSASSIN"
  | "KILLER_INSTINCT"
  | "SWISS_ARMY_KNIFE"
  | "TECHNICIAN"
  | "IRON_ENGINE"
  | "BULLDOG"
  | "SCRAPPER";